+++ lxblog/src/main/java/com/rahel/lxblog/dao/ArticleDao.java
@@ -1,6 +1,5 @@
 package com.rahel.lxblog.dao;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -12,20 +11,19 @@ import com.rahel.lxblog.entity.Article;
 import com.rahel.lxblog.entity.Tag_Article;
 
 @Repository
-public interface ArticleDao extends JpaRepository<Article, Tag_Article>{  //must be Integer
+public interface ArticleDao extends JpaRepository<Article, Tag_Article> { // must be Integer
 
-	ArrayList<Article> findAllByStatus(String status);
-	
+	List<Article> findAllByStatus(String status);
+
 	@Query("SELECT a FROM Article a where author_id=?1")
-	ArrayList<Article> findAllByAuthor_id(Integer author_id);
-	
-	List<Article> findAll();	
-	
+	List<Article> findAllByAuthor_id(Integer author_id);
+
+	List<Article> findAll();
+
 	Article save(Article article);
-	
+
 	Optional<Article> findById(Integer article_id);
-	
-	Integer deleteById(Integer id);
-	
-	
+
+	Integer deleteById(Integer id); // Shall I add annotation?
+
 }
