+++ lxblog/src/main/java/com/rahel/lxblog/model/ArticleRequest.java
@@ -6,27 +6,15 @@ public class ArticleRequest {
 
 	private String title;
 
-	private String text; 
+	private String text;
 
 	private String status;
-	
+
 	private List<String> tags;
-	
+
 	public ArticleRequest() {
-		
+
 	}
-	
-	/*public ArticleModel(Article article, List<String> tags) {
-		this.id = article.getId();
-		this.title=article.getTitle();
-		this.text=article.getText();
-		this.status=article.getStatus();
-		this.created_at= new java.util.Date(article.getCreated_at().getTime());// toLocalDate()); to edit this shit
-		if(article.getUpdated_at()!=null) {
-			this.updated_at= new java.util.Date(article.getUpdated_at().getTime());
-		}
-		this.tags=tags;
-	}*/
 
 	public String getTitle() {
 		return title;
